[#15] 피드 수정 기능 구현 - 권한이 없는 유저의 접근 차단 로직 추가

+++ src/main/java/com/ht/project/snsproject/service/FeedServiceImpl.java
@@ -179,7 +179,10 @@ public class FeedServiceImpl implements FeedService{
         builder.recommend(feedUpdateParam.getRecommend());
         FeedInsert feedInsert = builder.build();
 
-        feedMapper.updateFeed(feedInsert);
+        boolean result = feedMapper.updateFeed(feedInsert);
+        if(!result){
+            throw new InvalidApproachException("일치하는 데이터가 없습니다.");
+        }
         fileService.deleteFile(feedId);
         fileService.fileUpload(files, userId, feedId);
     }
